policy_service/requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.23
pydantic==1.10.15
python-dotenv>=1.0.1
llama-cpp-python>=0.2.86
